+++ .projectlintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  codemod: {
-    'lint-config-to-spec': 'error',
-    'plugin-rax-component-to-component': 'error',
-  }
-}